+++ ErsatzTV.Core/Iptv/ChannelPlaylist.cs
@@ -42,7 +42,7 @@ namespace ErsatzTV.Core.Iptv
 
                 string format = channel.StreamingMode switch
                 {
-                    StreamingMode.HttpLiveStreaming => "m3u8",
+                    StreamingMode.HttpLiveStreamingDirect => "m3u8",
                     _ => "ts"
                 };