+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+- Support `(Part #)` name suffixes for multi-part episode grouping
 
 ## [0.0.42-prealpha] - 2021-05-31
 ### Added