+++ CHANGELOG.md
@@ -4,7 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
 - Support `(Part #)` name suffixes for multi-part episode grouping
+- Support multi-episode files in local libraries
+
+### Fixed
+- Fix search result crashes due to missing season metadata
 
 ## [0.0.42-prealpha] - 2021-05-31
 ### Added