+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - Support `(Part #)` name suffixes for multi-part episode grouping
 - Support multi-episode files in local libraries
+- Save Channels table page size
 
 ### Fixed
 - Fix search result crashes due to missing season metadata